paddleocr-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
paddleocr>=2.9.0
paddlepaddle>=2.6.0
python-multipart>=0.0.6
pdf2image>=1.17.0
Pillow>=10.0.0
numpy<2.0.0
